Fix lint issues
Create SliderHandle component
