+++ src/components/Slider/Slider.jsx
@@ -1,152 +1,29 @@
-import React, { useCallback, useMemo, useRef } from "react";
+import React, { useCallback, useRef } from "react";
 import PropTypes from "prop-types";
 import {
   getClientPosition,
   addEventListeners,
   DIMENSIONS_MAP,
-  POSITIONS_MAP,
-  POSITIONS_MAP_CAPITALIZED
+  POSITIONS_MAP
 } from "./helpers";
 import { classNames } from "../../utils";
 import styles from "./styles.css";
-
-function useSliderAux(
-  container,
-  double,
-  disabled,
-  axis,
-  min,
-  max,
-  step,
-  value,
-  onChange,
-  onDragStart,
-  onDragEnd,
-  barrier
-) {
-  const handle = useRef(null);
-  const start = useRef({});
-  const offset = useRef({});
-
-  const getPosition = useMemo(() => {
-    let newValue = ((value - min) / (max - min)) * 100;
-
-    if (newValue > 100) newValue = 100;
-    if (newValue < 0) newValue = 0;
-
-    return axis === "x"
-      ? { top: 0, left: newValue }
-      : { top: newValue, left: 0 };
-  }, [value, min, max, axis]);
-
-  const change = useCallback(
-    (position) => {
-      if (!onChange) return;
-
-      const dimension = container.current.getBoundingClientRect()[
-        DIMENSIONS_MAP[axis]
-      ];
-      let ds = 0;
-
-      if (position < 0) position = 0;
-      if (position > dimension) position = dimension;
-
-      ds = (position / dimension) * (max - min);
-
-      const newPosition = (ds !== 0 ? parseInt(ds / step, 10) * step : 0) + min;
-
-      if (!double || barrier(newPosition)) {
-        onChange(newPosition);
-      }
-    },
-    [onChange, axis, min, max, step, double, barrier]
-  );
-
-  const getPos = useCallback((e) => {
-    const clientPos = getClientPosition(e);
-    const position =
-      clientPos[axis] + start.current[axis] - offset.current[axis];
-    return position;
-  }, []);
-
-  const handleDrag = useCallback(
-    (e) => {
-      if (disabled) return;
-
-      e.preventDefault();
-      change(getPos(e));
-    },
-    [disabled, change, getPos]
-  );
-
-  const handleDragEnd = useCallback(
-    (e) => {
-      if (disabled) return;
-
-      e.preventDefault();
-      document.removeEventListener("mousemove", handleDrag);
-      document.removeEventListener("touchmove", handleDrag, {
-        passive: false
-      });
-
-      document.removeEventListener("mouseup", handleDragEnd);
-      document.removeEventListener("touchend", handleDragEnd);
-      document.removeEventListener("touchcancel", handleDragEnd);
-
-      if (onDragEnd) {
-        onDragEnd(e);
-      }
-    },
-    [disabled, onDragEnd, handleDrag]
-  );
-
-  const handleMouseDown = useCallback(
-    (e) => {
-      if (disabled) return;
-
-      e.preventDefault();
-      e.stopPropagation();
-      e.nativeEvent.stopImmediatePropagation();
-
-      const dom = handle.current;
-      const clientPos = getClientPosition(e);
-
-      start.current[axis] = dom[`offset${POSITIONS_MAP_CAPITALIZED[axis]}`];
-      offset.current[axis] = clientPos[axis];
-
-      addEventListeners(handleDrag, handleDragEnd);
-
-      if (onDragStart) {
-        onDragStart(e);
-      }
-    },
-    [disabled, onDragStart, handleDrag, handleDragEnd]
-  );
-
-  return {
-    getPosition,
-    handleMouseDown,
-    handle,
-    handleDrag,
-    handleDragEnd,
-    change
-  };
-}
+import SliderHandle, { useSliderHandle } from "./SliderHandle";
 
 function useSlider(double, disabled, axis, min, max, step, handles) {
   const container = useRef(null);
 
   if (double) {
     handles[0].barrier = useCallback((value) => value <= handles[1].value, [
-      handles[1].value
+      handles
     ]);
     handles[1].barrier = useCallback((value) => value >= handles[0].value, [
-      handles[0].value
+      handles
     ]);
   }
 
-  const thumbs = handles.map((handle) =>
-    useSliderAux(
+  const handleHooks = handles.map((handle) =>
+    useSliderHandle(
       container,
       double,
       disabled,
@@ -170,48 +47,74 @@ function useSlider(double, disabled, axis, min, max, step, handles) {
       const clientPos = getClientPosition(e);
       const rect = container.current.getBoundingClientRect();
 
-      let moveSecondThumb = 0;
+      let handleIndex = 0;
 
       if (double) {
-        const thumbPositions = thumbs.map(
-          (thumb) => thumb.handle.current.getBoundingClientRect()[axis] + 9
+        const handlePosition = handleHooks.map(
+          (handleHook) =>
+            handleHook.handle.current.getBoundingClientRect()[axis] + 9
         );
 
         if (
-          (clientPos[axis] > thumbPositions[0] &&
-            clientPos[axis] < thumbPositions[1] &&
-            clientPos[axis] - thumbPositions[0] >=
-              thumbPositions[1] - clientPos[axis]) ||
-          clientPos[axis] >= thumbPositions[1]
+          (clientPos[axis] > handlePosition[0] &&
+            clientPos[axis] < handlePosition[1] &&
+            clientPos[axis] - handlePosition[0] >=
+              handlePosition[1] - clientPos[axis]) ||
+          clientPos[axis] >= handlePosition[1]
         )
-          moveSecondThumb = 1;
+          handleIndex = 1;
       }
 
       addEventListeners(
-        thumbs[moveSecondThumb].handleDrag,
-        thumbs[moveSecondThumb].handleDragEnd
+        handleHooks[handleIndex].handleDrag,
+        handleHooks[handleIndex].handleDragEnd
       );
 
-      thumbs[moveSecondThumb].change(
+      handleHooks[handleIndex].change(
         clientPos[axis] - rect[POSITIONS_MAP[axis]]
       );
 
-      if (thumbs[moveSecondThumb].onDragStart) {
-        thumbs[moveSecondThumb].onDragStart(e);
+      if (handleHooks[handleIndex].onDragStart) {
+        handleHooks[handleIndex].onDragStart(e);
       }
     },
-    [axis, disabled, double, thumbs]
+    [axis, disabled, double, handleHooks]
   );
+
+  const sliderHandles = handleHooks.map((handleHook, index) => {
+    const onPositionMemo = handleHook.positionMemo;
+    return (
+      <SliderHandle
+        key={index}
+        ref={handleHook.handle}
+        axis={axis}
+        positionMemo={onPositionMemo}
+        onTouchStart={handleHook.handleMouseDown}
+        onMouseDown={handleHook.handleMouseDown}
+      />
+    );
+  });
+
+  const valueStyle = {};
+
+  if (double) {
+    valueStyle[POSITIONS_MAP[axis]] =
+      handleHooks[0].positionMemo[POSITIONS_MAP[axis]] + "%";
+
+    valueStyle[DIMENSIONS_MAP[axis]] =
+      handleHooks[1].positionMemo[POSITIONS_MAP[axis]] -
+      handleHooks[0].positionMemo[POSITIONS_MAP[axis]] +
+      "%";
+  } else {
+    valueStyle[DIMENSIONS_MAP[axis]] =
+      handleHooks[0].positionMemo[POSITIONS_MAP[axis]] + "%";
+  }
 
   return {
     container,
     handleTrackMouseDown,
-    getPosition: thumbs[0].getPosition,
-    handleMouseDown: thumbs[0].handleMouseDown,
-    handle: thumbs[0].handle,
-    getPosition2: thumbs[1]?.getPosition,
-    handleMouseDown2: thumbs[1]?.handleMouseDown,
-    handle2: thumbs[1]?.handle
+    valueStyle,
+    sliderHandles
   };
 }
 
@@ -228,32 +131,9 @@ const Slider = ({
   const {
     container,
     handleTrackMouseDown,
-    getPosition,
-    handleMouseDown,
-    handle,
-    getPosition2,
-    handleMouseDown2,
-    handle2
+    valueStyle,
+    sliderHandles
   } = useSlider(double, disabled, axis, min, max, step, handles);
-
-  const pos = getPosition;
-
-  const pos2 = getPosition2;
-
-  const valueStyle = {};
-  const handleStyle = { top: "50%", left: "50%" };
-  const handleStyle2 = { top: "50%", left: "50%" };
-
-  handleStyle[POSITIONS_MAP[axis]] = pos[POSITIONS_MAP[axis]] + "%";
-
-  if (!double) {
-    valueStyle[DIMENSIONS_MAP[axis]] = pos[POSITIONS_MAP[axis]] + "%";
-  } else {
-    valueStyle[POSITIONS_MAP[axis]] = pos[POSITIONS_MAP[axis]] + "%";
-    valueStyle[DIMENSIONS_MAP[axis]] =
-      pos2[POSITIONS_MAP[axis]] - pos[POSITIONS_MAP[axis]] + "%";
-    handleStyle2[POSITIONS_MAP[axis]] = pos2[POSITIONS_MAP[axis]] + "%";
-  }
 
   return (
     <div
@@ -266,36 +146,12 @@ const Slider = ({
       )}
       onTouchStart={handleTrackMouseDown}
       onMouseDown={handleTrackMouseDown}>
-      {double && (
-        <div
-          ref={handle2}
-          style={handleStyle2}
-          className={styles.handle}
-          onTouchStart={handleMouseDown2}
-          onMouseDown={handleMouseDown2}
-          onClick={(e) => {
-            e.stopPropagation();
-            e.nativeEvent.stopImmediatePropagation();
-          }}>
-          <div className={styles.thumb} />
-        </div>
-      )}
+      {double && sliderHandles[1]}
       <div
         className={classNames(styles[axis], styles.active)}
         style={valueStyle}
       />
-      <div
-        ref={handle}
-        style={handleStyle}
-        className={styles.handle}
-        onTouchStart={handleMouseDown}
-        onMouseDown={handleMouseDown}
-        onClick={(e) => {
-          e.stopPropagation();
-          e.nativeEvent.stopImmediatePropagation();
-        }}>
-        <div className={styles.thumb} />
-      </div>
+      {sliderHandles[0]}
     </div>
   );
 };
